+++ api/lib/infrastructure/plugins/index.js
@@ -1,0 +1,16 @@
+const config = require('../../config');
+const Blipp = require('blipp');
+const Inert = require('@hapi/inert');
+const Vision = require('@hapi/vision');
+
+const plugins = [
+  require('./metrics'),
+  Inert,
+  Vision,
+  Blipp,
+  require('./adminjs'),
+  require('./pino'),
+  ...(config.sentry.enabled ? [require('./sentry')] : []),
+];
+
+module.exports = plugins;
