+++ api/lib/infrastructure/plugins/sentry.js
@@ -1,0 +1,19 @@
+const Pack = require('../../../package');
+const config = require('../../config');
+
+module.exports = {
+  plugin: require('hapi-sentry'),
+  options: {
+    client: {
+      dsn: config.sentry.dsn,
+      environment: config.sentry.environment,
+      release: `v${Pack.version}`,
+      maxBreadcrumbs: config.sentry.maxBreadcrumbs,
+      debug: config.sentry.debug,
+      maxValueLength: config.sentry.maxValueLength,
+    },
+    scope: {
+      tags: [{ name: 'source', value: 'api' }],
+    },
+  },
+};
