+++ api/server.js
@@ -7,7 +7,7 @@ const Oppsy = require('@hapi/oppsy');
 const preResponseUtils = require('./lib/infrastructure/utils/pre-response-utils');
 
 const routes = require('./lib/routes');
-const plugins = require('./lib/plugins');
+const plugins = require('./lib/infrastructure/plugins');
 const config = require('./lib/config');
 const security = require('./lib/infrastructure/security');
 const securityPreHandlers = require('./lib/application/security-pre-handlers');
